Cherry-picks for the 3.2 branch (future 3.2.4) - 7th batch

+++ platform/windows/joypad_windows.h
@@ -78,7 +78,7 @@ private:
 		DWORD last_pad;
 
 		LPDIRECTINPUTDEVICE8 di_joy;
-		List<DWORD> joy_axis;
+		List<LONG> joy_axis;
 		GUID guid;
 
 		dinput_gamepad() {
